Slight Rework of H5BP.org (#30)
This PR changes the project to use babel, which is nice enough for now, but will help in the future. 

Speaking of the future, we're still using $ because I didn't want to be _that_ ambitious. Next time. 

The bigger change is to hide repos that have been archived. So the list is slightly shorter but won't show dead ends. 

Now, to develop, run `npm run babel` and then browse the directory. I've just been using `serve` on the directory. That could be improved as well.